+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Event: `event:character.prompt`
 - Event: `event:menu.choose`
 - Some utility functions for `Persistent` and `Transform`
+- Use a handler to set persistent data
 
 ### Fixed